changes done

+++ database/migrations/2023_06_12_174201_create_purchases_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->timestamp('date', $precision = 0);
             $table->text('desc')->nullable();
             $table->string('isPaid');
-            $table->unsignedBigInteger('amount')->nullable();
+            $table->float('amount')->nullable();
             $table->unsignedBigInteger('ref');
             $table->foreign('vendor')->references('id')->on('accounts');
             $table->foreign('paidFrom')->references('id')->on('accounts');
