+++ database/migrations/2023_06_15_155604_create_sales_table.php
@@ -19,9 +19,9 @@ return new class extends Migration
             $table->timestamp('date', $precision = 0);
             $table->text('desc')->nullable();
             $table->string('isPaid');
-            $table->unsignedBigInteger('amount')->nullable();
-            $table->unsignedInteger('discount')->nullable();
-            $table->unsignedBigInteger('ref');
+            $table->float('amount')->nullable();
+            $table->float('discount')->nullable();
+            $table->integer('ref');
             $table->foreign('customer')->references('id')->on('accounts');
             $table->foreign('paidIn')->references('id')->on('accounts');
             $table->timestamps();
