+++ database/migrations/2023_06_18_171234_create_ledgers_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('head');
             $table->string('type');
             $table->string('details');
-            $table->unsignedBigInteger('amount');
+            $table->float('amount');
             $table->unsignedBigInteger('ref');
             $table->timestamps();
         });
